Update the db user grant to include password
I continue to see the user of the database being generated without the database password being set correctly and I believe I have tracked it down to the database cookbook which allows you to create users and grant user permissions.

The grant operation needs to have the password provided otherwise it alters the user to have a blank password.